FetchAPINews/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
mysql-connector-python==8.0.26
protobuf==3.19.0
requests==2.26.0
urllib3==1.26.7
